fix some small bugs realted to cgi post

+++ src/Client/Client.hpp
@@ -62,7 +62,7 @@ class Client {
 		bool			fileCreated;
 		bool			canIStart;
 		bool			hasCgi;
-		int				totalBytesRead;
+		long				totalBytesRead;
 		int				Content_Length;
 		bool			_readHeader;
 		serverBlock		*_serverBlock;
